pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aura-state"
version = "0.1.0"
description = "A Neurosymbolic, Self-Healing State Router for Production LLMs."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.0",
    "instructor>=1.3",
    "openai>=1.0",
    "python-dotenv>=1.0.0",
    "networkx>=3.0",
    "pyyaml>=6.0",
    "pyModelChecking>=1.3",
    "z3-solver>=4.12",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio"
]
